Review fixes: colors module, translatable statuses, schema split, preview-layer + polling optimizations
- Named colors in dialogs/colors.py; ProcessingView uses them.
- Template/AOI table_status and labels go through QCoreApplication.translate.
- Split schema/processing.py -> schema/template.py (one-directional imports).
- Preview fixes: remove stale mosaic boundary layer; de-duplicate image previews
  (move existing to top); in-template layer precedence AOIs > previews > search.
- Always send searchParams.aoiDetails on template creation (plain aoi deprecated);
  names optional, single unnamed feature for image/mosaic extents.
- Polling: in-template tick is a single /processings request (no get_template/images),
  slower cadence; table rebuild blocks selection signals so it no longer re-triggers
  the AOI search filter / map rebuild; AOI status synced from polled processings.
