+++ mapflow/functional/api/processing_api.py
@@ -9,6 +9,8 @@ from ...schema.processing import (
     PostProcessingSchema,
     UpdateProcessingSchema,
     ProcessingsRequest,
+)
+from ...schema.template import (
     CreateProcessingTemplateSchema,
     UpdateProcessingTemplateSchema,
     RunTemplateProcessingSchema,
